+++ packages/Textarea/src/Textarea.tsx
@@ -3,6 +3,9 @@ import cx from 'classnames';
 import autosize from 'autosize';
 import { mergeRefs } from '@react-aria/utils';
 
+import useCharLength from './hooks/useCharLength';
+import useTruncateValue from './hooks/useTruncateValue';
+
 import './textarea.scss';
 
 export interface TextareaProps extends React.ComponentPropsWithRef<'textarea'> {
@@ -54,13 +57,12 @@ const Textarea: React.FunctionComponent<TextareaProps> = React.forwardRef(
   ) => {
     const textareaRef = React.useRef<HTMLTextAreaElement | null>(null);
     const mergedTextareaRef = mergeRefs(textareaRef, ref);
-    const [currentCharLength, setCurrentCharLength] = React.useState(
-      value?.length ?? 0
-    );
     const [currentValue, setCurrentValue] = React.useState(value ?? '');
     const textareaMaxLength = maxLength ?? 0;
+    const charLength = useCharLength(currentValue, textareaMaxLength);
     const displayCharIndicator =
       showCharactersIndicator && textareaMaxLength > 0;
+    const truncateValue = useTruncateValue();
 
     const classes = cx('ids-textarea', className, {
       'ids-textarea--has-char-count': displayCharIndicator,
@@ -74,7 +76,8 @@ const Textarea: React.FunctionComponent<TextareaProps> = React.forwardRef(
     const handleOnChange = (
       event: React.ChangeEvent<HTMLTextAreaElement>
     ): void => {
-      setCurrentValue(event.target.value);
+      const newValue = truncateValue(event.target.value, maxLength);
+      setCurrentValue(newValue);
       if (onChange) {
         onChange(event);
       }
@@ -106,18 +109,9 @@ const Textarea: React.FunctionComponent<TextareaProps> = React.forwardRef(
     }, [textareaRef, isAutoResize]);
 
     React.useEffect(() => {
-      let currentValueLength = currentValue?.length ?? 0;
-
-      const emojiRegexExp =
-        /(\u00a9|\u00ae|[\u2000-\u3300]|\ud83c[\ud000-\udfff]|\ud83d[\ud000-\udfff]|\ud83e[\ud000-\udfff])/gi;
-      const hasEmoji = emojiRegexExp.test(currentValue);
-
-      if (currentValue.length === maxLength! - 1 && hasEmoji) {
-        currentValueLength = currentValue.length + 1;
-      }
-
-      setCurrentCharLength(currentValueLength);
-    }, [currentValue]);
+      const newValue = truncateValue(value?.toString() ?? '', maxLength);
+      setCurrentValue(newValue);
+    }, [value, maxLength, truncateValue]);
 
     return (
       <div className={classes} data-test={dataTest}>
@@ -127,14 +121,14 @@ const Textarea: React.FunctionComponent<TextareaProps> = React.forwardRef(
           maxLength={maxLength}
           onChange={handleOnChange}
           onKeyDown={handleKeyDown}
-          value={value}
+          value={currentValue}
           disabled={disabled}
           {...rest}
         />
 
         {displayCharIndicator && (
           <div className="ids-textarea__char-indicator">
-            {textareaMaxLength - currentCharLength}
+            {textareaMaxLength - charLength}
           </div>
         )}
       </div>